+++ Repositories/RepositoryMesa.py
@@ -1,0 +1,4 @@
+from Repositories.InterfaceRepository import InterfaceRepository
+from Models.Mesa import Mesa
+class RepositoryMesa(InterfaceRepository[Mesa]):
+  pass
