--- Repositories/RepositoryVotingTable.py
@@ -1,4 +1,0 @@
-from Repositories.InterfaceRepository import InterfaceRepository
-from Models.VotingTable import VotingTable
-class RepositoryVotingTable(InterfaceRepository[VotingTable]):
-  pass
