+++ main.py
@@ -19,7 +19,7 @@ print(baseDatos.list_collection_names())
 app=Flask(__name__)
 cors = CORS(app)
 
-from Controllers.ControllerVotingTable import ControllerVotingTable
+from Controllers.ControllerMesa import ControllerVotingTable
 myControllerVotingTable = ControllerVotingTable()
 
 @app.route("/votingtables",methods=['GET'])
